+++ odoo/local-src/schweizmobil_subs_bundle_ee/__manifest__.py
@@ -20,7 +20,7 @@
         'sale_subscription_to_close_next_date_advance',
         # local-src
         'schweizmobil_subs_sale_subscription',
-        'schweizmobil_subs_sale_subscription_renewal'
+        'schweizmobil_subs_sale_subscription_renewal',
     ],
     'installable': True,
     'auto_install': False,