+++ odoo/local-src/wim_connector/components/binder.py
@@ -4,6 +4,5 @@ from odoo.addons.component.core import Component
 
 
 class WimBinder(Component):
-
     _name = "wim.binder"
     _inherit = ["wim.base", "base.binder"]