+++ odoo/local-src/wim_connector/models/__init__.py
@@ -1,4 +1,4 @@
 from . import wim_backend
 from . import wim_binding
 from . import res_partner
-from . import sale_subscription
+from . import sale_order