+++ odoo/songs/migration_db/cleanup_config.py
@@ -34,7 +34,6 @@ UNINSTALL_MODULES_LIST = [
     # local-src
     'schweizmobil_account_bank_statement_import',
     'schweizmobil_report',
-    'wim_connector',
     'wim_rest',
 ]